[Feat]#7 - viewModel 하나로 합치기
